Added File Events and made UglifyJS listen to that (#50)
* Added File Event

* Made ugilify listen to File event instead of Asset event

* Fixed event names

* Codestyle

* Really correct the names

+++ test/FileSystem/FileWriterTest.php
@@ -5,8 +5,15 @@
 declare(strict_types=1);
 namespace Hostnet\Component\Resolver\FileSystem;
 
+use Hostnet\Component\Resolver\Event\FileEvent;
+use Hostnet\Component\Resolver\Event\FileEvents;
 use Hostnet\Component\Resolver\File;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Argument;
+use Symfony\Component\EventDispatcher\Event;
+use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
  * @covers \Hostnet\Component\Resolver\FileSystem\FileWriter
@@ -15,7 +22,12 @@ class FileWriterTest extends TestCase
 {
     public function testGeneric()
     {
-        $writer = new FileWriter(__DIR__);
+        $dispatcher = $this->prophesize(EventDispatcherInterface::class);
+
+        $dispatcher->dispatch(FileEvents::PRE_WRITE, Argument::type(FileEvent::class))->shouldBeCalled();
+        $dispatcher->dispatch(FileEvents::POST_WRITE, Argument::type(FileEvent::class))->shouldBeCalled();
+
+        $writer = new FileWriter($dispatcher->reveal(), __DIR__);
         $writer->write(new File('output/foobar.txt'), 'foobar');
 
         self::assertSame('foobar', file_get_contents(__DIR__ . '/output/foobar.txt'));
@@ -24,10 +36,25 @@ class FileWriterTest extends TestCase
         self::assertSame('baz', file_get_contents(__DIR__ . '/output/baz.txt'));
     }
 
+    public function testChangeContentWithListener()
+    {
+        $dispatcher = new EventDispatcher();
+        $dispatcher->addListener(FileEvents::PRE_WRITE, function (FileEvent $e) {
+            $e->setContent('barbaz');
+        });
+
+        $writer = new FileWriter($dispatcher, __DIR__);
+        $writer->write(new File('output/foobar.txt'), 'foobar');
+
+        self::assertSame('barbaz', file_get_contents(__DIR__ . '/output/foobar.txt'));
+    }
+
     protected function tearDown()
     {
         unlink(__DIR__ . '/output/foobar.txt');
-        unlink(__DIR__ . '/output/baz.txt');
+        if (file_exists(__DIR__ . '/output/baz.txt')) {
+            unlink(__DIR__ . '/output/baz.txt');
+        }
         rmdir(__DIR__ . '/output');
     }
 }
